+++ desktop/cmp/treemap/TreeMapModel.js
@@ -8,7 +8,7 @@ import {HoistModel} from '@xh/hoist/core';
 import {action, bindable, computed, observable, makeObservable} from '@xh/hoist/mobx';
 import {throwIf, withDefault} from '@xh/hoist/utils/js';
 import {numberRenderer} from '@xh/hoist/format';
-import {cloneDeep, get, isEmpty, isFinite, max, mean, set, sumBy, unset} from 'lodash';
+import {cloneDeep, get, isEmpty, isFinite, max, set, sumBy, unset} from 'lodash';
 
 /**
  * Core Model for a TreeMap.
@@ -23,7 +23,6 @@ import {cloneDeep, get, isEmpty, isFinite, max, mean, set, sumBy, unset} from 'l
  *
  * Node colors are normalized to a 0-1 range and mapped to a colorAxis via the following colorModes:
  * 'linear' distributes normalized color values across the colorAxis according to the heatField.
- * 'balanced' attempts to account for outliers by adjusting normalisation ranges around the median.
  * 'wash' ignores the intensity of the heat value, applying a single positive and negative color.
  * 'none' will ignore the colorAxis, and instead use the neutral color.
  *
@@ -102,7 +101,7 @@ export class TreeMapModel extends HoistModel {
      * @param {string} [c.algorithm] - Layout algorithm to use. Either 'squarified',
      *     'sliceAndDice', 'stripes' or 'strip'. Defaults to 'squarified'.
      *     {@see https://www.highcharts.com/docs/chart-and-series-types/treemap} for examples.
-     * @param {string} [c.colorMode] - Heat color distribution mode. Either 'linear', 'balanced', 'wash' or
+     * @param {string} [c.colorMode] - Heat color distribution mode. Either 'linear', 'wash' or
      *     'none'. Defaults to 'linear'.
      * @param {function} [c.onClick] - Callback to call when a node is clicked. Receives (record,
      *     e). If not provided, by default will select a record when using a GridModel.
@@ -153,7 +152,7 @@ export class TreeMapModel extends HoistModel {
         throwIf(!['sliceAndDice', 'stripes', 'squarified', 'strip'].includes(algorithm), `Algorithm "${algorithm}" not recognised.`);
         this.algorithm = algorithm;
 
-        throwIf(!['linear', 'balanced', 'wash', 'none'].includes(colorMode), `Color mode "${colorMode}" not recognised.`);
+        throwIf(!['linear', 'wash', 'none'].includes(colorMode), `Color mode "${colorMode}" not recognised.`);
         this.colorMode = colorMode;
 
         this.onClick = withDefault(onClick, this.defaultOnClick);
@@ -292,11 +291,8 @@ export class TreeMapModel extends HoistModel {
      *
      * Takes the colorMode into account:
      * a) 'linear' distributes normalized color values across the colorAxis.
-     * b) 'balanced' attempts to account for outliers by adjusting the normalisation ranges around
-     *  the median values. Can result in more defined color differences in a dataset that is skewed
-     *  by a few nodes at the extremes.
-     * c) 'wash' ignores the intensity of the heat value, applying a flat positive or negative color.
-     * d) 'none' ignores the heat value altogether, coloring all nodes with the neutral color
+     * b) 'wash' ignores the intensity of the heat value, applying a flat positive or negative color.
+     * c) 'none' ignores the heat value altogether, coloring all nodes with the neutral color
      */
     normaliseColorValues(data) {
         const {colorMode, heatField} = this;
@@ -324,9 +320,9 @@ export class TreeMapModel extends HoistModel {
             return data;
         }
 
-        //---------------------
-        // ColorMode === 'linear|balanced'
-        //---------------------
+        //------------------------
+        // ColorMode === 'linear'
+        //------------------------
         // 1) Extract valid heat values
         const heatValues = [];
         this.store.records.forEach(it => {
@@ -335,10 +331,7 @@ export class TreeMapModel extends HoistModel {
         });
 
         // 2) Transform heatValue into a normalized colorValue, according to the colorMode.
-        const minHeat = 0,
-            midHeat = mean(heatValues),
-            maxHeat = isFinite(this.maxHeat) ? this.maxHeat : max(heatValues);
-
+        const maxHeat = isFinite(this.maxHeat) ? this.maxHeat : max(heatValues);
         data.forEach(it => {
             const {heatValue} = it;
 
@@ -349,27 +342,10 @@ export class TreeMapModel extends HoistModel {
 
             if (heatValue > 0) {
                 // Normalize positive values between 0.6-1
-                if (colorMode === 'balanced' && heatValues.length > 2) {
-                    if (it.colorValue >= midHeat) {
-                        it.colorValue = this.normalizeToRange(heatValue, midHeat, maxHeat, 0.8, 1);
-                    } else {
-                        it.colorValue = this.normalizeToRange(heatValue, minHeat, midHeat, 0.6, 0.8);
-                    }
-                } else if (colorMode === 'linear' || heatValues.length <= 2) {
-                    it.colorValue = this.normalizeToRange(heatValue, minHeat, maxHeat, 0.6, 1);
-                }
+                it.colorValue = this.normalizeToRange(heatValue, 0, maxHeat, 0.6, 1);
             } else if (heatValue < 0) {
                 // Normalize negative values between 0-0.4
-                const absHeatValue = Math.abs(heatValue);
-                if (colorMode === 'balanced' && heatValues.length > 2) {
-                    if (absHeatValue >= midHeat) {
-                        it.colorValue = this.normalizeToRange(absHeatValue, maxHeat, midHeat, 0, 0.2);
-                    } else {
-                        it.colorValue = this.normalizeToRange(absHeatValue, midHeat, minHeat, 0.2, 0.4);
-                    }
-                } else if (colorMode === 'linear' || heatValues.length <= 2) {
-                    it.colorValue = this.normalizeToRange(absHeatValue, maxHeat, minHeat, 0, 0.4);
-                }
+                it.colorValue = this.normalizeToRange(Math.abs(heatValue), maxHeat, 0, 0, 0.4);
             } else {
                 it.colorValue = 0.5; // Exactly zero
             }
@@ -378,21 +354,11 @@ export class TreeMapModel extends HoistModel {
         return data;
     }
 
-    /**
-     * Takes a value, calculates its normalized (0-1) value within a specified range.
-     * If a destination range is provided, returns that value transposed to within that range.
-     */
     normalizeToRange(value, fromMin, fromMax, toMin, toMax) {
         const fromRange = (fromMax - fromMin),
             toRange = (toMax - toMin);
 
-        if (isFinite(toRange)) {
-            // Return value transposed to destination range
-            return (((value - fromMin) * toRange) / fromRange) + toMin;
-        } else {
-            // Return value between 0-1
-            return (value - fromMin) / fromRange;
-        }
+        return (((value - fromMin) * toRange) / fromRange) + toMin;
     }
 
     valueIsValid(value) {